requirements.txt
pathlib
textwrap
google.generativeai
gradio
pymongo
